Split up in api.jugendstadtplan.de and www.jugendstadtplan.de
Now the Project has a separate subdomain api.jugendstadtplan.de/dev for the backend and a subdomain www.jugendstadtplan.de/dev for the frontend. The backend is a regular Symfony app, the frontend is a Angular single page app.
In this setup CORS is enabled, so that the frontend can request all infos from the backend without hassle. This also works with the "production" environment (before this it didn't).